requirements.txt
adafruit_ads1x15==1.0.2
Adafruit_Blinka==6.16.0
adafruit_bme280==1.0.1
adafruit_ccs811==0.2.1
adafruit_circuitpython_ads1x15==2.2.10
adafruit_circuitpython_bme280==2.6.7
adafruit_circuitpython_ccs811==1.3.5
AWSIoTPythonSDK==1.4.9
board==1.0
python-dotenv==0.19.2
